+++ packages/ui/select/select-option.module.css.d.ts
@@ -1,5 +1,6 @@
 declare const styles: {
   readonly "select-option": string;
+  readonly "defult": string;
   readonly "option": string;
 };
 export = styles;