+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,43 @@
 # cosmic-design
 
+## 0.0.21
+
+### Patch Changes
+
+-   [`4f158a4`](https://github.com/design-to-release/cosmic-design/commit/4f158a4b396e836de8cb09ccc3615ccb64571d3b) Thanks [@qiansc](https://github.com/qiansc)! - update color set
+
+*   [`344f390`](https://github.com/design-to-release/cosmic-design/commit/344f3900bc66dd1041b5129b655f5de428ac9308) Thanks [@qiansc](https://github.com/qiansc)! - update color set
+
+-   [`f5c7602`](https://github.com/design-to-release/cosmic-design/commit/f5c7602b485d10661dc68d77c925b13830d0cae4) Thanks [@qiansc](https://github.com/qiansc)! - update color set
+
+## 0.0.20
+
+### Patch Changes
+
+-   [#80](https://github.com/design-to-release/cosmic-design/pull/80) [`0dde61d`](https://github.com/design-to-release/cosmic-design/commit/0dde61d25cd5aa38328e4c1c8f720d0981ba5215) Thanks [@biyingshuai](https://github.com/biyingshuai)! - update tree module css
+
+## 0.0.19
+
+### Patch Changes
+
+-   [`0002a31`](https://github.com/design-to-release/cosmic-design/commit/0002a311902e7f1f1c3a6c9644c87888c5dbe36f) Thanks [@qiansc](https://github.com/qiansc)! - update version
+
+*   [`bdea5c1`](https://github.com/design-to-release/cosmic-design/commit/bdea5c1b5294db503516f6b451b780770dd3d15f) Thanks [@qiansc](https://github.com/qiansc)! - update version
+
+## 0.0.18
+
+### Patch Changes
+
+-   [`5dd9332`](https://github.com/design-to-release/cosmic-design/commit/5dd93322e6c67b3ff361b97630522bf51d60fa08) Thanks [@qiansc](https://github.com/qiansc)! - update css order
+
+## 0.0.17
+
+### Patch Changes
+
+-   [`8259ff6`](https://github.com/design-to-release/cosmic-design/commit/8259ff63cd562add2215d0fdde0f678bfab3ad3a) Thanks [@qiansc](https://github.com/qiansc)! - fix: fix icon size && add button module css
+
+*   [`862d7ff`](https://github.com/design-to-release/cosmic-design/commit/862d7ff9ee13d4e696af255cfe011e39c9172149) Thanks [@qiansc](https://github.com/qiansc)! - upgrade ui
+
 ## 0.0.16
 
 ### Patch Changes